js/xpconnect/tests/chrome/test_bug658560.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=658560
-->
<window title="Mozilla Bug 658560"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=658560"
     target="_blank">Mozilla Bug 658560</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

      SimpleTest.waitForExplicitFinish();
      function go() {
        var win = $('ifr').contentWindow.wrappedJSObject;
        let o = { prop: true };
        win.foo = o;

        is(win.foo, o, "should have === identity through a CrossOriginWrapper");
        SimpleTest.finish();
      }

  ]]></script>

  <iframe
      src="http://example.org/tests/js/xpconnect/tests/mochitest/file_bug658560.html"
          id="ifr"
          type="content"
          onload="go()" />
</window>
